Python_online/website_button_press/requirements.txt
selenium
webdriver_manager.microsoft